pyproject.toml
[project]
name = "de-streaming-data"
version = "0.1.0"
description = "Lambda function to retrieve guardian articles containing query"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.37.29",
    "httpx>=0.28.1",
    "moto>=5.1.3",
    "pytest-cov>=6.1.1",
    "pytest-testdox>=3.1.0",
    "python-dotenv>=1.1.0",
    "respx>=0.22.0",
    "ruff>=0.11.4",
]

[tool.ruff]
line-length = 80
indent-width = 4

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 70

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B"]

[tool.pytest.ini_options]
addopts = "--testdox --cov-report term-missing --cov=src/"
filterwarnings = ["ignore::DeprecationWarning"]
